requirements.txt
discord.py
requests
langdetect